feat: move serialize_value_as_json to common
