+++ crates/cheatcodes/src/json.rs
@@ -4,7 +4,7 @@ use crate::{Cheatcode, Cheatcodes, Result, Vm::*, string};
 use alloy_dyn_abi::{DynSolType, DynSolValue, Resolver, eip712_parser::EncodeType};
 use alloy_primitives::{Address, B256, I256, hex};
 use alloy_sol_types::SolValue;
-use foundry_common::fs;
+use foundry_common::{fmt::serialize_value_as_json, fs};
 use foundry_config::fs_permissions::FsAccessKind;
 use serde_json::{Map, Value};
 use std::{borrow::Cow, collections::BTreeMap};
@@ -582,49 +582,6 @@ pub(super) fn json_value_to_token(value: &Value) -> Result<DynSolValue> {
             }
             Ok(DynSolValue::String(string.to_owned()))
         }
-    }
-}
-
-/// Serializes given [DynSolValue] into a [serde_json::Value].
-fn serialize_value_as_json(value: DynSolValue) -> Result<Value> {
-    match value {
-        DynSolValue::Bool(b) => Ok(Value::Bool(b)),
-        DynSolValue::String(s) => {
-            // Strings are allowed to contain stringified JSON objects, so we try to parse it like
-            // one first.
-            if let Ok(map) = serde_json::from_str(&s) {
-                Ok(Value::Object(map))
-            } else {
-                Ok(Value::String(s))
-            }
-        }
-        DynSolValue::Bytes(b) => Ok(Value::String(hex::encode_prefixed(b))),
-        DynSolValue::FixedBytes(b, size) => Ok(Value::String(hex::encode_prefixed(&b[..size]))),
-        DynSolValue::Int(i, _) => {
-            // let serde handle number parsing
-            let n = serde_json::from_str(&i.to_string())?;
-            Ok(Value::Number(n))
-        }
-        DynSolValue::Uint(i, _) => {
-            // let serde handle number parsing
-            let n = serde_json::from_str(&i.to_string())?;
-            Ok(Value::Number(n))
-        }
-        DynSolValue::Address(a) => Ok(Value::String(a.to_string())),
-        DynSolValue::Array(e) | DynSolValue::FixedArray(e) => {
-            Ok(Value::Array(e.into_iter().map(serialize_value_as_json).collect::<Result<_>>()?))
-        }
-        DynSolValue::CustomStruct { name: _, prop_names, tuple } => {
-            let values =
-                tuple.into_iter().map(serialize_value_as_json).collect::<Result<Vec<_>>>()?;
-            let map = prop_names.into_iter().zip(values).collect();
-
-            Ok(Value::Object(map))
-        }
-        DynSolValue::Tuple(values) => Ok(Value::Array(
-            values.into_iter().map(serialize_value_as_json).collect::<Result<_>>()?,
-        )),
-        DynSolValue::Function(_) => bail!("cannot serialize function pointer"),
     }
 }
