+++ crates/common/fmt/src/lib.rs
@@ -7,7 +7,8 @@ pub use console::{ConsoleFmt, FormatSpec, console_format};
 
 mod dynamic;
 pub use dynamic::{
-    format_token, format_token_raw, format_tokens, format_tokens_raw, parse_tokens, token_to_json,
+    format_token, format_token_raw, format_tokens, format_tokens_raw, parse_tokens,
+    serialize_value_as_json,
 };
 
 mod exp;
